+++ CHANGELOG.md
@@ -16,6 +16,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Autocrafting now handles multiple patterns with the same output correctly by trying to use the pattern with the
     highest priority first. If there are missing resources, lower priority patterns are checked.
 
+### Fixed
+
+-   Fixed amount in amount screens resetting when resizing the screen.
+
 ## [2.0.0-milestone.4.11] - 2024-12-08
 
 ### Added